pyproject.toml
[tool.poetry]
name = "task_exchange_system"
version = "0.1.0"
description = "Task Exchange System project"
authors = []

[tool.poetry.dependencies]
python = "~3.10"
fastapi = "0.95.2"
uvicorn = {extras = ["standart"], version = "0.22.0"}
pydantic = {extras = ["email", "dotenv"], version = "1.10.8"}
psycopg2-binary = "2.9.6"
sqlalchemy = "2.0.15"
alembic = "1.11.1"
asyncpg = "0.27.0"
greenlet = "2.0.2"
httpx = "0.24.1"

JSON-log-formatter = "0.3.1"
python-json-logger = "2.0.7"
python-dateutil = "2.8.2"
fastapi-sso = "^0.7.2"
aiokafka = "^0.8.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
pylama = {extras = ["toml", "pylint", "pycodestyle", "pyflakes"], version = "8.4.1"}
isort = "5.12.0"
black = "23.3.0"
pytest = "7.3.1"
pytest-asyncio = "0.21.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylama]
linters = "pylint,pycodestyle,pyflakes"
max_line_length = 119
skip = "**/migrations/*"
format = "pylint"
ignore = "C0114,C0115,C0116,C0103,C0121,C0411,R0903,W0107,W1203,W0511,E0611,E0213,E0401"

[[tool.pylama.files]]
path = "**/__init__.py"
ignore = "W0611"

[tool.black]
line-length = 119
target-version = ['py310']
extend-exclude = "/migrations/"

[tool.isort]
py_version = 310
profile = "black"
line_length = 119
skip_glob = ["**/migrations/*", ]
# здесь должны быть перечислены все директории проекта
known_local_folder = [
    "application",
    "controllers",
    "db",
    "dto",
    "middleware",
    "migrations",
    "repositories",
    "services",
]
